feat: integration HomeStack
